backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-Cors==4.0.0
PyMySQL==1.0.3
python-dotenv==1.0.0
pytest==7.4.3
gunicorn==21.2.0
Flask-JWT-Extended==4.5.2
